updating checkAttr
